Algunas correcciones y Clase InscripcionData
Se generan metodos en claseInscripcionData: guardarInscripcion.
Se corrige el tipo de dato del atributo nota, de inscripcion, de int a double.

+++ src/universidadgrupo49/Entidades/Inscripcion.java
@@ -6,18 +6,18 @@ public class Inscripcion {
     private int idInscripcion;
     private Alumno alumno;
     private Materia materia;
-    private int nota;
+    private double nota;
 
     public Inscripcion() {
     }
 
-    public Inscripcion(Alumno alumno, Materia materia, int nota) {
+    public Inscripcion(Alumno alumno, Materia materia, double nota) {
         this.alumno = alumno;
         this.materia = materia;
         this.nota = nota;
     }
 
-    public Inscripcion(int idInscripcion, Alumno alumno, Materia materia, int nota) {
+    public Inscripcion(int idInscripcion, Alumno alumno, Materia materia, double nota) {
         this.idInscripcion = idInscripcion;
         this.alumno = alumno;
         this.materia = materia;
@@ -48,11 +48,11 @@ public class Inscripcion {
         this.materia = materia;
     }
 
-    public int getNota() {
+    public double getNota() {
         return nota;
     }
 
-    public void setNota(int nota) {
+    public void setNota(double nota) {
         this.nota = nota;
     }
